+++ Src/Metrics.Tests/Sampling/UniformSnapshotTests.cs
@@ -1,4 +1,5 @@
-﻿using FluentAssertions;
+﻿using System;
+using FluentAssertions;
 using Metrics.Sampling;
 using Xunit;
 
@@ -6,7 +7,7 @@ namespace Metrics.Tests.Sampling
 {
     public class UniformSnapshotTest
     {
-        private readonly UniformSnapshot snapshot = new UniformSnapshot(new long[] { 5, 1, 2, 3, 4 });
+        private readonly UniformSnapshot snapshot = new UniformSnapshot(5, new long[] { 5, 1, 2, 3, 4 });
 
         [Fact]
         public void UniformSnapshot_SmallQuantilesAreTheFirstValue()
@@ -95,36 +96,44 @@ namespace Metrics.Tests.Sampling
         [Fact]
         public void UniformSnapshot_CalculatesAMinOfZeroForAnEmptySnapshot()
         {
-            Snapshot snapshot = new UniformSnapshot(new long[] { });
+            Snapshot snapshot = new UniformSnapshot(0, new long[] { });
             snapshot.Min.Should().Be(0);
         }
 
         [Fact]
         public void UniformSnapshot_CalculatesAMaxOfZeroForAnEmptySnapshot()
         {
-            Snapshot snapshot = new UniformSnapshot(new long[] { });
+            Snapshot snapshot = new UniformSnapshot(0, new long[] { });
             snapshot.Max.Should().Be(0);
         }
 
         [Fact]
         public void UniformSnapshot_CalculatesAMeanOfZeroForAnEmptySnapshot()
         {
-            Snapshot snapshot = new UniformSnapshot(new long[] { });
+            Snapshot snapshot = new UniformSnapshot(0, new long[] { });
             snapshot.Mean.Should().Be(0);
         }
 
         [Fact]
         public void UniformSnapshot_CalculatesAStdDevOfZeroForAnEmptySnapshot()
         {
-            Snapshot snapshot = new UniformSnapshot(new long[] { });
+            Snapshot snapshot = new UniformSnapshot(0, new long[] { });
             snapshot.StdDev.Should().Be(0);
         }
 
         [Fact]
         public void UniformSnapshot_CalculatesAStdDevOfZeroForASingletonSnapshot()
         {
-            Snapshot snapshot = new UniformSnapshot(new long[] { 1 });
+            Snapshot snapshot = new UniformSnapshot(0, new long[] { 1 });
             snapshot.StdDev.Should().Be(0);
+        }
+
+        [Fact]
+        public void UniformSnapshot_ThrowsOnBadQuantileValue()
+        {
+            ((Action)(() => snapshot.GetValue(-0.5))).ShouldThrow<ArgumentException>();
+            ((Action)(() => snapshot.GetValue(1.5))).ShouldThrow<ArgumentException>();
+            ((Action)(() => snapshot.GetValue(double.NaN))).ShouldThrow<ArgumentException>();
         }
     }
 }