+++ Src/Metrics/Json/JsonProperty.cs
@@ -12,6 +12,14 @@ namespace Metrics.Json
             : this(name, new ObjectJsonValue(properties))
         { }
 
+        public JsonProperty(string name, JsonObject @object)
+            : this(name, new ObjectJsonValue(@object))
+        { }
+
+        public JsonProperty(string name, IEnumerable<string> value)
+            : this(name, new StringArrayJsonValue(value))
+        { }
+
         public JsonProperty(string name, string value)
             : this(name, new StringJsonValue(value))
         { }