+++ Src/Metrics/PerfCounters/PerformanceCounterGauge.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Diagnostics;
 using System.Security.Principal;
+using Metrics.MetricData;
 
 namespace Metrics.PerfCounters
 {
@@ -19,6 +20,7 @@ namespace Metrics.PerfCounters
                 this.performanceCounter = instance == null ?
                     new PerformanceCounter(category, counter, true) :
                     new PerformanceCounter(category, counter, instance, true);
+                Metric.Internal.Counter("Performance Counters", Unit.Custom("Perf Counters")).Increment();
             }
             catch (Exception x)
             {
@@ -38,6 +40,11 @@ namespace Metrics.PerfCounters
             {
                 return "[Unknown user | " + x.Message + " ]";
             }
+        }
+
+        public double GetValue(bool resetMetric = false)
+        {
+            return this.Value;
         }
 
         public double Value