+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+###0.2.10 / 2014-11-06
+* fix error logging for not found performance counters
+
+###0.2.9 / 2014-11-05
+* record active sessions for timers
+
+###0.2.8 / 2014-10-29
+* handle access issues for perf counters
+
+###0.2.7 / 2014-10-28
+* preparations for out-of-process metrics
+
+###0.2.6 / 2014-10-17
+* fix http listener prefix handling
+
+###0.2.5 / 2014-10-12
+* JSON metrics refactor
+* remote metrics 
+
 ###0.2.4 / 2014-10-07
 * JSON version
 * added environment